backendFlask/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
PyJWT==2.7.0
bcrypt==4.0.1